fix: pass pagespeed insights
