--- src/routes/service_worker.rs
@@ -1,25 +1,0 @@
-use axum::{http::header, response::IntoResponse};
-
-use crate::template::Template;
-
-#[derive(askama::Template)]
-#[template(path = "sw.js")]
-pub struct ServiceWorkerTemplate;
-
-/// Serve service worker at /sw.js (must be at root for scope: '/')
-pub async fn asset(template: Template) -> impl IntoResponse {
-    (
-        [
-            (
-                header::CONTENT_TYPE.as_str(),
-                "application/javascript; charset=utf-8",
-            ),
-            (
-                header::CACHE_CONTROL.as_str(),
-                "no-cache, no-store, must-revalidate",
-            ),
-            ("Service-Worker-Allowed", "/"),
-        ],
-        template.render(ServiceWorkerTemplate),
-    )
-}
