+++ templates/_base.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
-<html lang="{{ " html_lang"|t }}">
+<html lang="{{ "html_lang"|t }}">
 
 <head>
   <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta name="description"
     content="{% block meta_description %}Intelligent meal planning and cooking optimization{% endblock %}">
