+++ templates/_footer.html
@@ -3,7 +3,7 @@
   <div class="container mx-auto px-4">
     <div class="grid md:grid-cols-4 gap-8 mb-8">
       <div>
-        <div class="text-2xl font-bold text-white mb-4">🍳 imkitchen</div>
+        <div class="text-2xl font-bold text-white mb-4"><span aria-hidden="true">🍳</span> imkitchen</div>
         <p class="text-sm text-gray-400">{{ "Intelligent meal planning for home cooks."|t }}</p>
       </div>
       <div>
