+++ templates/robots.txt
@@ -1,0 +1,5 @@
+User-agent: *
+Allow: /
+Disallow: /static/
+
+Sitemap: https://imkitchen.app/sitemap.xml
