+++ templates/sitemap.xml
@@ -1,0 +1,35 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url>
+    <loc>https://imkitchen.app/</loc>
+    <priority>1.0</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/about</loc>
+    <priority>0.7</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/help</loc>
+    <priority>0.5</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/contact</loc>
+    <priority>0.5</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/terms</loc>
+    <priority>0.3</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/policy</loc>
+    <priority>0.3</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/register</loc>
+    <priority>0.8</priority>
+  </url>
+  <url>
+    <loc>https://imkitchen.app/login</loc>
+    <priority>0.6</priority>
+  </url>
+</urlset>
